Feedback implementation 1

+++ client/src/js/actions/groupAction.js
@@ -1,4 +1,3 @@
-import jwt from 'jsonwebtoken';
 import Api from '../utils/api';
 import {
   ADD_GROUP_FAIL,
@@ -9,9 +8,6 @@ import {
   REMOVE_CURRENT_GROUP,
   RECEIVE_MESSAGES,
   RECEIVE_MESSAGES_SUCCESS,
-  RECEIVE_MESSAGES_FAIL,
-  ADD_USER_SUCCESS,
-  ADD_USER_FAIL,
   GET_GROUP_MEMBERS_SUCCESS,
   GET_GROUP_MEMBERS_FAIL,
   REMOVE_GROUP_MEMBER_SUCCESS,
@@ -21,8 +17,8 @@ import {
 
 
 /**
- * @param {*} void
- * @return {*} empty
+ * @param {null} void
+ * @return {object} type
  */
 export function getGroup() {
   return {
@@ -31,8 +27,8 @@ export function getGroup() {
 }
 
 /**
- * @param {*} userGroups
- * @return {*} empty
+ * @param {object} userGroups
+ * @return {object} type & userGroups
  */
 export function getGroupSuccess(userGroups) {
   return {
@@ -42,8 +38,8 @@ export function getGroupSuccess(userGroups) {
 }
 
 /**
- * @param {*} groupError
- * @return {*} empty
+ * @param {object} groupError
+ * @return {object} type & groupError
  */
 export function getGroupFail(groupError) {
   return {
@@ -53,8 +49,8 @@ export function getGroupFail(groupError) {
 }
 
 /**
- * @param {*} currentGroup
- * @return {*} empty
+ * @param {object} currentGroup
+ * @return {object} type and currentGroup
  */
 export function setCurrentGroup(currentGroup) {
   return {
@@ -64,8 +60,8 @@ export function setCurrentGroup(currentGroup) {
 }
 
 /**
- * @param {*} currentGroupMembers
- * @return {*} empty
+ * @param {object} currentGroupMembers
+ * @return {object} empty
  */
 export function setCurrentGroupUsersSuccess(currentGroupMembers) {
   return {
@@ -75,8 +71,8 @@ export function setCurrentGroupUsersSuccess(currentGroupMembers) {
 }
 
 /**
- * @param {*} groupMembersError
- * @return {*} empty
+ * @param {object} groupMembersError
+ * @return {object} empty
  */
 export function setCurrentGroupUsersFail(groupMembersError) {
   return {
@@ -88,7 +84,7 @@ export function setCurrentGroupUsersFail(groupMembersError) {
 
 /**
  * @param {*} void
- * @return {*} empty
+ * @return {object} empty
  */
 export function removeCurrentGroup() {
   return {
@@ -170,7 +166,9 @@ export function removeGroupMember(userId) {
     userId
   };
 }
-
+/**
+ * @return {void}
+ */
 export const fetchUserGroups = () => (dispatch) => {
   dispatch(getGroup());
   // making call to the get all groups API endpoint
@@ -183,6 +181,12 @@ export const fetchUserGroups = () => (dispatch) => {
   });
 };
 
+/**
+ * 
+ * @param {*} name 
+ * @param {*} desc
+ * @return {void}
+ */
 export const createGroup = (name, desc) => (dispatch) => {
   const groupCreateParams = `name=${name}&desc=${desc}`;
   // making calls to the create group API endpoint
@@ -195,12 +199,19 @@ export const createGroup = (name, desc) => (dispatch) => {
     }
   ).catch((addError) => {
     dispatch(addGroupFail(addError));
+    throw addError;
   });
 };
 
 
   // onSearchUserInGroup Method
-export const onSearchUser = (id, searchText) => dispatch => Api(
+  /**
+   * 
+   * @param {*} id 
+   * @param {*} searchText
+   * @return {void}
+   */
+export const onSearchUser = (id, searchText) => () => Api(
   null,
   `/api/v1/groups/${id}/usersearch?search=${searchText}`,
   'GET')
@@ -209,9 +220,14 @@ export const onSearchUser = (id, searchText) => dispatch => Api(
   );
 
   // TODO: fix users in a group and deleting user from group UI & backend
-export const setSelectedGroupMembers = groupId => (dispatch) => {
+  /**
+   * 
+   * @param {*} groupId
+   * @return {void}
+   */
+export const setSelectedGroupMembers = groupId => dispatch =>
   // TODO: make API call to get users of a group from the server
-  return Api(null, `/api/v1/groups/${groupId}/users/`, 'GET').then(
+  Api(null, `/api/v1/groups/${groupId}/users/`, 'GET').then(
     (currentGroupMembers) => {
       // setting current group members response from server to store
       dispatch(setCurrentGroupUsersSuccess(currentGroupMembers));
@@ -222,51 +238,73 @@ export const setSelectedGroupMembers = groupId => (dispatch) => {
       // setting error response from server when getting members of group
       dispatch(setCurrentGroupUsersFail(groupMembersError));
       return groupMembersError;
-    });
-};
+    })
+;
 
-  // onAddUserToGroup Method
-export const onAddUser = (userId, groupId) => (dispatch) => {
+// onAddUserToGroup Method
+/**
+   * 
+   * @param {*} userId 
+   * @param {*} groupId
+   * @return {object} addUserToGroupResponse
+   */
+export const onAddUser = (userId, groupId) => () => {
   const addUserParams = `userId=${userId}`;
   return Api(addUserParams, `/api/v1/groups/${groupId}/user/`, 'POST').then(
-    (addUserToGroupResponse) => {
-      // check if response, i.e user added, set message state to "user added"
-      // check for users in a group and update UI to add mark icon
-      return addUserToGroupResponse;
-    }
-
+    addUserToGroupResponse => addUserToGroupResponse
   );
 };
 
 // onAddUserToGroup Method
+/**
+ * 
+ * @param {*} userId 
+ * @param {*} groupId 
+ * @return {object} removeUserFromGroupResponse
+ */
 export const onRemoveUser = (userId, groupId) => (dispatch) => {
   const removeUserParams = `?usersId=${userId}`;
-  return Api(null, `/api/v1/groups/${groupId}/user/${removeUserParams}`, 'DELETE').then(
-    (removeUserFromGroupResponse) => {
-      console.log('removeUserFromGroupResponse========>', removeUserFromGroupResponse);
-      if (removeUserFromGroupResponse.message === 'User Is not an Admin') {
+  return Api(
+    null,
+    `/api/v1/groups/${groupId}/user/${removeUserParams}`,
+    'DELETE')
+    .then(
+      (removeUserFromGroupResponse) => {
+        if (removeUserFromGroupResponse.message ===
+          'User Is not an Admin' || 'You cannot remove yourself') {
+          return removeUserFromGroupResponse;
+        }
+        // call remove groupmember action
+        dispatch(removeGroupMember(userId));
         return removeUserFromGroupResponse;
       }
-      if (removeUserFromGroupResponse.message === 'You cannot remove yourself') {
-        return removeUserFromGroupResponse;
-      }
-      // call remove groupmember action
-      dispatch(removeGroupMember(userId));
-      return removeUserFromGroupResponse;
-    }
-  );
+    );
 };
 
+/**
+ * 
+ * @param {*} group
+ * @return {void}
+ */
 export const setSelectedGroupAsCurrent = group => (dispatch) => {
   // setting current group to store
   dispatch(setCurrentGroup(group));
 };
 
+/**
+ * @return {void}
+ */
 export const resetCurrentGroup = () => (dispatch) => {
   // setting current group to store
   dispatch(removeCurrentGroup());
 };
 
+/**
+ * 
+ * @param {*} groupId 
+ * @param {*} dispatch 
+ * @return {object} groupMessages
+ */
 const fetchTheMessages = (groupId, dispatch) => {
   dispatch(getGroupMessages());
   return Api(null, `/api/v1/groups/${groupId}/messages/`, 'GET').then(
@@ -277,8 +315,21 @@ const fetchTheMessages = (groupId, dispatch) => {
   );
 };
 
-export const fetchMessages = groupId => dispatch => fetchTheMessages(groupId, dispatch);
+/**
+ * 
+ * @param {*} groupId
+ * @return {object} messageResponse
+ */
+export const fetchMessages = groupId => dispatch =>
+  fetchTheMessages(groupId, dispatch);
 
+/**
+ * 
+ * @param {*} groupId 
+ * @param {*} priority 
+ * @param {*} text 
+ * @return {object} messageResponse
+ */
 export const handleSendMessage = (groupId, priority, text) => (dispatch) => {
   dispatch(addMessage());
   // make Api call to send msg here
@@ -290,17 +341,5 @@ export const handleSendMessage = (groupId, priority, text) => (dispatch) => {
       return messageResponse;
     });
 };
-// TODO: fix this and abstract from messageboard component
-// export const updateReadBy = id => (dispatch) => {
-//   const updateMessageParams = `id=${id}`;
-//   const gId = `${this.props.match.params.groupId}`;
-//   // make API call to the read message endpoint
-//   Api(updateMessageParams, `/api/v1/groups/${gId}/message/read`, 'POST')
-//     .then((response) => {
-//       console.log('Response: ', response);
-//     });
-// };
-
-// TODO: fix responsiveness css
 
 // TODO: abstract all other API calls in component
